+++ yVAR.h
@@ -159,8 +159,8 @@ yVAR_expect        (void);
 int                /*  return  = pos=good, neg=bad, 0=unknown                 */
 yVAR_integer(      /*  PURPOSE = STANDARD INTEGER TESTING (for yUNIT+)        */
       char     *a_test,          /*  name of the test                         */
-      long      a_expect,        /*  expected result                          */
-      long      a_actual)        /*  actual result                            */
+      long long a_expect,        /*  expected result                          */
+      long long a_actual)        /*  actual result                            */
 ;
 
 int                /*  return  = pos=good, neg=bad, 0=unknown                 */
